+++ qom/solvers/__init__.py
@@ -1,1 +1,3 @@
+from qom.solvers.HLESolver import HLESolver
+from qom.solvers.QCMSolver import QCMSolver
 from qom.solvers.RHCSolver import RHCSolver